Fix quoting of values in POST_RECOVERY_SCRIPT and BACKUP_PROG_INCLUDE
Add code to write correctly quoted values in the POST_RECOVERY_SCRIPT
BACKUP_PROG_INCLUDE variables in the ReaR configuration file and to read
them back. The values were previously output only with escaped spaces.
For instance, this caused a problem for the snapper quota script as it
contains various special characters. This then resulted in the following
error when running ReaR (bsc#1180983):
> /etc/rear/local.conf: line 18: syntax error near unexpected token '|'

The module newly distinguishes between two types of lists that are used
in the ReaR configuration. They get called in the code as basic and
quoted lists:
* A basic list consists of as-is names. Elements of this list are not
  transformed in any way when being written to or read from the
  configuration file, with the exception of escaping spaces.
* A quoted list consists of names that should be protected from shell
  processing. Elements of such a list are quoted when writing them in
  the configuration file and unquoted/unescaped when reading them. This
  is used to store scripts that are invoked by ReaR using
  'eval "${CONFIG_VAR[@]}"' and for wildcard names that should not be
  expanded by the shell directly.

The already present support for basic lists remains mostly unmodified.
Implementation methods RearListToYCPList() and YCPListToRearList() are
only modernized to use native Ruby operations instead of the Builtins
and Ops methods.

Methods RearQuotedListToYCPList() and YCPListToRearQuotedList() are
newly introduced to handle quoted lists. Variables POST_RECOVERY_SCRIPT
and BACKUP_PROG_INCLUDE are switched to use these methods.

Update functionality is then added to offer the user to replace an
incorrectly escaped version of the snapper quota script in
POST_RECOVERY_SCRIPT that was produced by the module previously.